refactor: split 4 oversized modules into focused submodules [OPE-259, OPE-260, OPE-261]
- schedule.rs (761 LOC) → schedule/{mod,commands,logic,tests}.rs
- dashboard.rs (871 LOC) → dashboard/{mod,activity,metrics,tests}.rs
- views/mod.rs (891 LOC) → views/{mod,dashboard,queue,runs,schedules,shared,status,tests,triggers,workflows}.rs
- openapi/schemas.rs (656 LOC) → schemas/{mod,agents,alerts,channels,dashboard,events,gateways,responses,runs,sessions,system,teams,tests,triggers,workflows}.rs
- Fix pub(crate) visibility for session_store re-export
- All 2,096 tests pass, clean clippy
