+++ crates/opengoose-web/src/data/views/runs.rs
@@ -1,0 +1,57 @@
+use super::shared::MetaRow;
+
+/// Summary row for the orchestration run list sidebar.
+#[derive(Clone)]
+pub struct RunListItem {
+    pub title: String,
+    pub subtitle: String,
+    pub updated_at: String,
+    pub progress_label: String,
+    pub badge: String,
+    pub badge_tone: &'static str,
+    pub page_url: String,
+    pub queue_page_url: String,
+    pub active: bool,
+}
+
+/// A single work item row in the run detail panel.
+#[derive(Clone)]
+pub struct WorkItemView {
+    pub title: String,
+    pub detail: String,
+    pub status_label: String,
+    pub status_tone: &'static str,
+    pub step_label: String,
+    pub indent_class: &'static str,
+}
+
+/// A broadcast message shown in the run detail panel.
+#[derive(Clone)]
+pub struct BroadcastView {
+    pub sender: String,
+    pub created_at: String,
+    pub content: String,
+}
+
+/// Full detail panel for a selected orchestration run.
+#[derive(Clone)]
+pub struct RunDetailView {
+    pub title: String,
+    pub subtitle: String,
+    pub source_label: String,
+    pub meta: Vec<MetaRow>,
+    pub work_items: Vec<WorkItemView>,
+    pub broadcasts: Vec<BroadcastView>,
+    pub input: String,
+    pub result: String,
+    pub empty_hint: String,
+}
+
+/// View-model for the runs page (list + selected detail).
+#[derive(Clone)]
+pub struct RunsPageView {
+    pub mode_label: String,
+    pub mode_tone: &'static str,
+    pub runs: Vec<RunListItem>,
+    pub selected: RunDetailView,
+}
