+++ crates/opengoose-web/src/data/views/shared.rs
@@ -1,0 +1,81 @@
+/// A single metric card rendered on the dashboard (label, value, footnote, tone).
+#[derive(Clone)]
+pub struct MetricCard {
+    pub label: String,
+    pub value: String,
+    pub note: String,
+    pub tone: &'static str,
+}
+
+/// An alert banner displayed on the dashboard.
+#[derive(Clone)]
+pub struct AlertCard {
+    pub eyebrow: String,
+    pub title: String,
+    pub description: String,
+    pub tone: &'static str,
+}
+
+/// One segment of a stacked status bar (e.g. "Running 3" at 40% width).
+#[allow(dead_code)]
+#[derive(Clone)]
+pub struct StatusSegment {
+    pub label: String,
+    pub value: String,
+    pub tone: &'static str,
+    pub width: u8,
+}
+
+/// A single bar in the duration trend chart.
+#[allow(dead_code)]
+#[derive(Clone)]
+pub struct TrendBar {
+    pub label: String,
+    pub value: String,
+    pub detail: String,
+    pub tone: &'static str,
+    pub height: u8,
+}
+
+/// One row in the activity feed timeline.
+#[allow(dead_code)]
+#[derive(Clone)]
+pub struct ActivityItem {
+    pub actor: String,
+    pub meta: String,
+    pub detail: String,
+    pub timestamp: String,
+    pub tone: &'static str,
+}
+
+/// A label/value metadata row shown in detail panels.
+#[derive(Clone)]
+pub struct MetaRow {
+    pub label: String,
+    pub value: String,
+}
+
+/// Option row for a `<select>` field.
+#[derive(Clone)]
+pub struct SelectOption {
+    pub value: String,
+    pub label: String,
+    pub selected: bool,
+}
+
+/// A toast-style notice shown after an action (e.g. team save).
+#[derive(Clone)]
+pub struct Notice {
+    pub text: String,
+    pub tone: &'static str,
+}
+
+/// A gateway connection status card for the dashboard widget.
+#[derive(Clone)]
+pub struct GatewayCard {
+    pub platform: String,
+    pub state_label: String,
+    pub state_tone: &'static str,
+    pub uptime_label: String,
+    pub detail: String,
+}
