+++ crates/opengoose-web/src/openapi/schemas/agents.rs
@@ -1,0 +1,13 @@
+use serde_json::{Value, json};
+
+pub(super) fn build() -> Value {
+    json!({
+        "AgentItem": {
+            "type": "object",
+            "required": ["name"],
+            "properties": {
+                "name": { "type": "string", "description": "Profile name" }
+            }
+        }
+    })
+}
